Add files via upload
In this notebook, I explore sentiment classification on Twitter data, which contains tweets labeled with either positive or negative sentiment. The base model I utilized is a pre-trained SentenceTransformer model called all-MiniLM-L6-v2. This baseline model accurately classifies the sentiment of tweets within the dataset 77% of the time.

The goal of this analysis is to enhance the model's ability to accurately classify the sentiment of tweets by fine-tuning the model's embeddings to improve its capacity to distinguish between positive and negative sentiments.